+++ src/Stormancer.Plugins/Notification/Stormancer.Server.Plugins.Notification/INotificationProvider.cs
@@ -23,12 +23,23 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Stormancer.Server.Plugins.Notification
 {
+    /// <summary>
+    /// Provides API to send notifications.
+    /// </summary>
     public interface INotificationProvider
     {
-        Task<bool> SendNotification(string type, dynamic data);
+        /// <summary>
+        /// Sends a notification.
+        /// </summary>
+        /// <param name="type"></param>
+        /// <param name="data"></param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
+        Task<bool> SendNotification(string type, dynamic data,CancellationToken cancellationToken);
     }
 }