+++ src/Stormancer.Plugins/Stormancer.Server.TestApp/App.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Stormancer.Server.TestApp
-{
-    public class App
-    {
-    }
-}