+++ src/Stormancer.Plugins/DataProtection/Stormancer.Server.Plugins.DataProtection/Changelog.rst
@@ -10,9 +10,17 @@ Please use only reStructuredText in this file, no Markdown!
 This project adheres to semantic versioning.
 
 
+1.0.1.1
+-------
+Changed
+*******
+- Use Stormancer.Abstraction.Server 5.0.0
+
 1.0.1
 -----
-- Use Stormancer.Abstraction.Server 6.1.0
+Changed
+*******
+- Use Stormancer.Abstraction.Server 4.1.0
 
 1.0.0.1
 ----------