+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -9,6 +9,19 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+5.0.1.2
+-------
+Changed
+*******
+- Make plugin compatible with Users 5.x
+- Use Stormancer.Abstraction.Server 5.0.0
+
+5.0.0
+-----
+Removed
+*******
+- Removed deprecated OpenToGameFinder API. Use GameFinderProxy.OpenGameSession instead.
+
 4.0.7
 -----
 Changed