+++ src/Stormancer.Plugins/Utilities/Stormancer.Server.Plugins.Annotations/Changelog.rst
@@ -9,6 +9,12 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+1.1.0
+-----
+Added
+*****
+- Added GeneratePrivateImpl to S2SApiAttribute to force the code generator to generate the action proxy method as a private method of the proxy.
+
 1.0.0
 ------
 Added